Add files via upload
It is an Excel sheet that has detailed analysis of IMDB movies.The objective of the project is to find what are all the factors that make a movie successful.Here success of a movie is determined by its IMDB score. So , we need to find the factors that influencing IMDB score.